refactor(-): :hammer: move e remove arquivos inversao dia cardapio
